Add changed_on date field to changed prayers
Whenever a prayer has a change (as denoted by: "changed": true),
it will also have "changed_on": <iso-date>. This denotes the day
on which the change was detected. This is not a timestamp because
typically changes happen on midnight, but we only scrape them later
in the day.

This closes the backend for #10.